[agent] refactor(cli): move restore errors to error module
