+++ crates/gaze-cli/src/main.rs
@@ -14,7 +14,9 @@ use std::time::Duration;
 
 use base64::engine::general_purpose::STANDARD as BASE64;
 use base64::Engine;
-use clap::{Parser, Subcommand, ValueEnum};
+mod error;
+
+use clap::{Parser, Subcommand};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 
@@ -24,6 +26,8 @@ use gaze::{
     Scope, SensitiveSnapshot, Session,
 };
 use gaze_recognizers::{NerDetector, NerOptions, RegexDetector};
+
+use crate::error::{CliError, RestoreMode, RestoreWarning};
 
 #[derive(Parser, Debug)]
 #[command(
@@ -69,84 +73,6 @@ enum Cmd {
         #[arg(long, default_value_t = DEFAULT_MAX_BYTES)]
         max_bytes: u64,
     },
-}
-
-/// Structured CLI error. Each variant maps to an exit code; only the variant
-/// name reaches stderr so raw input or plaintext blob entries never leak into
-/// caller logs (see docs/roadmap/v0.3/cli.md "Stderr discipline").
-#[derive(Debug)]
-enum CliError {
-    StdinParse,
-    EmptyInput,
-    InputTooLarge,
-    InvalidEncoding,
-    PolicyConfig,
-    PolicyConfigDetail(&'static str),
-    UnknownToken { token: String },
-    InvalidSignature,
-    InvalidBlobVersion,
-    BlobExpired,
-    Pipeline,
-    Io,
-    PolicyOpen,
-}
-
-impl CliError {
-    fn exit_code(&self) -> u8 {
-        match self {
-            Self::StdinParse | Self::EmptyInput | Self::InputTooLarge | Self::InvalidEncoding => 1,
-            Self::PolicyConfig | Self::PolicyConfigDetail(_) => 2,
-            Self::UnknownToken { .. }
-            | Self::InvalidSignature
-            | Self::InvalidBlobVersion
-            | Self::BlobExpired
-            | Self::Pipeline => 3,
-            Self::Io | Self::PolicyOpen => 4,
-        }
-    }
-
-    fn variant_name(&self) -> &'static str {
-        match self {
-            Self::StdinParse => "StdinParse",
-            Self::EmptyInput => "EmptyInput",
-            Self::InputTooLarge => "InputTooLarge",
-            Self::InvalidEncoding => "InvalidEncoding",
-            Self::PolicyConfig | Self::PolicyConfigDetail(_) => "PolicyConfig",
-            Self::UnknownToken { .. } => "UnknownToken",
-            Self::InvalidSignature => "InvalidSignature",
-            Self::InvalidBlobVersion => "InvalidBlobVersion",
-            Self::BlobExpired => "BlobExpired",
-            Self::Pipeline => "Pipeline",
-            Self::Io => "Io",
-            Self::PolicyOpen => "PolicyOpen",
-        }
-    }
-
-    fn emit_stderr(&self) {
-        match self {
-            Self::PolicyConfigDetail(detail) => eprintln!(
-                r#"{{"error":"{}","exit":{},"detail":"{}"}}"#,
-                self.variant_name(),
-                self.exit_code(),
-                detail
-            ),
-            Self::UnknownToken { token } => {
-                let token = serde_json::to_string(token)
-                    .unwrap_or_else(|_| "\"<unserializable>\"".to_string());
-                eprintln!(
-                    r#"{{"error":"{}","exit":{},"token":{}}}"#,
-                    self.variant_name(),
-                    self.exit_code(),
-                    token
-                )
-            }
-            _ => eprintln!(
-                r#"{{"error":"{}","exit":{}}}"#,
-                self.variant_name(),
-                self.exit_code()
-            ),
-        }
-    }
 }
 
 /// Install a panic hook that prints a sanitized error line and exits 3.
@@ -325,12 +251,6 @@ fn run_clean(
     Ok(())
 }
 
-#[derive(ValueEnum, Clone, Copy, Debug, Eq, PartialEq)]
-enum RestoreMode {
-    Strict,
-    Tolerant,
-}
-
 fn run_restore(
     format: &str,
     restore_mode: RestoreMode,
@@ -404,7 +324,9 @@ fn build_pipeline_from_policy(policy: &Policy) -> GazeResult<Pipeline> {
 
     for rule in &policy.rules {
         builder = match rule {
-            RuleSpec::Class { class, action } => builder.rule(ClassRule::new(class.clone(), *action)),
+            RuleSpec::Class { class, action } => {
+                builder.rule(ClassRule::new(class.clone(), *action))
+            }
             RuleSpec::Column { column, action } => builder.rule(ColumnRule::new(column, *action)),
             RuleSpec::Default { action } => builder.rule(DefaultRule::new(*action)),
         };
@@ -513,12 +435,6 @@ struct RestoreResponse {
     text: String,
     #[serde(skip_serializing_if = "Vec::is_empty")]
     restore_warning: Vec<RestoreWarning>,
-}
-
-#[derive(Serialize)]
-struct RestoreWarning {
-    variant: String,
-    token: String,
 }
 
 /// Stub pipeline used until the policy.toml loader (solo #3) lands.
